Updated Mock + Added Typescript and ts-node
