+++ package.json
@@ -14,7 +14,7 @@
     "lint:fix": "tslint --fix -p ./",
     "ionic:build": "ionic-app-scripts build",
     "ionic:serve": "ionic-app-scripts serve",
-    "serve:local": "run-s copy-mock-data config:local ionic:serve",
+    "serve:local": "run-s update-mock-data copy-mock-data config:local ionic:serve",
     "ionic:deploy": "git push ionic master",
     "format:fix": "pretty-quick --staged",
     "format:all": "npx prettier --write \"src/**/*{.ts,.js,.json,.css,.scss}\"",
@@ -34,7 +34,7 @@
     "config:production": "cp src/environment/environment.production.ts src/environment/environment.ts",
     "config:local": "cp src/environment/environment.local.ts src/environment/environment.ts",
     "copy-mock-data": "cp -r mock/ src/assets/mock/",
-    "update-mock-data": "node bin/update-mock-data.js"
+    "update-mock-data": "ts-node bin/update-mock-data.ts"
   },
   "dependencies": {
     "@angular/animations": "5.2.11",
